feat: organize packages
