+++ requirements.txt
@@ -1,1 +1,6 @@
-pandas langchain langchain_community chromadb sentence_transformers langchain_huggingface
+pandas
+langchain
+langchain_community
+chromadb
+sentence_transformers
+langchain_huggingface
